ServerEmailPython/requirements.txt
aiosmtpd==1.4.6
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
python-dotenv==1.1.0